Update README.md
I had to redeploy my project on Vercel which caused the link to change. I redeployed because while testing the hosted project on my smartphone, I found out that visitors of the hosted site were required to login to Vercel. I have turned off the setting on my Vercel account and redeployed